SDX-1146 Moved files inside selections.
