+++ src/framework/filterControl/_index.scss
@@ -3,6 +3,6 @@
 @import 'filterDropdown/filterDropdown';
 @import 'filterOptions/FilterOptionList';
 @import 'filterSelections/conditionCheckerList';
-@import 'forms/filterForm';
-@import 'forms/multipleSelectFilterForm';
-@import 'forms/singleSelectFilterForm';
+@import 'filterSelections/forms/filterForm';
+@import 'filterSelections/forms/multipleSelectFilterForm';
+@import 'filterSelections/forms/singleSelectFilterForm';
