+++ src/framework/filterControl/filterSelections/forms/_multipleSelectFilterForm.scss
