+++ src/framework/filterControl/filterSelections/forms/_singleSelectFilterForm.scss
